Handle nested types
